chore: split WEB and CLI codebase
